Organise folder structure for the ofences panel
