+++ src/features/CourtCaseDetails/Tabs/Panels/Offences/OffencesList/OffencesList.tsx
@@ -1,39 +1,15 @@
 import { Offence } from "@moj-bichard7-developers/bichard7-next-core/build/src/types/AnnotatedHearingOutcome"
 import { Table } from "govuk-react"
-import { formatDisplayedDate } from "utils/formattedDate"
-import getOffenceCode from "utils/getOffenceCode"
 
-interface OffencesTableProps {
+import getOffenceCode from "utils/getOffenceCode"
+import { OffencesListRow } from "./OffencesListItem"
+
+interface OffencesListProps {
   offences: Offence[]
   setDetailedOffence: (offence: Offence) => void
 }
 
-interface HearingTableRowProps {
-  offence: Offence
-  number: number
-  onClick: (offence: Offence) => void
-}
-
-const HearingTableRow = ({ offence, onClick, number }: HearingTableRowProps) => (
-  <Table.Row>
-    <Table.Cell>{number}</Table.Cell>
-    <Table.Cell>{formatDisplayedDate(offence.ActualOffenceStartDate.StartDate)}</Table.Cell>
-    <Table.Cell>{getOffenceCode(offence)}</Table.Cell>
-    <Table.Cell>
-      <a
-        href="#"
-        onClick={(e) => {
-          e.preventDefault()
-          onClick(offence)
-        }}
-      >
-        {offence.OffenceTitle}
-      </a>
-    </Table.Cell>
-  </Table.Row>
-)
-
-export const OffencesTable = ({ offences, setDetailedOffence }: OffencesTableProps) => {
+export const OffencesList = ({ offences, setDetailedOffence }: OffencesListProps) => {
   return (
     <Table
       head={
@@ -47,7 +23,7 @@ export const OffencesTable = ({ offences, setDetailedOffence }: OffencesTablePro
     >
       {offences.length > 0 &&
         offences.map((offence, index) => (
-          <HearingTableRow
+          <OffencesListRow
             key={getOffenceCode(offence)}
             offence={offence}
             number={index + 1}
